Add files via upload
TASK 3:  Atm Interface
In this project I showcase how can we create ATM interface that can interact with user and ask for their choices which  are offer in the banks.
User need to provide first his user account and pin no. which is  then verified from the bank's own information and if it is correct then can only use avail the services.